requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
anthropic
xai_grok_sdk
xai-sdk
gunicorn
grpcio
grpcio-tools
pymongo[srv]